Undo works for polygon was drawn by PolygonTool
